Fixed another Integer index problem

+++ BlockLocalNMF_AuxilaryFunctions.py
@@ -473,7 +473,7 @@ def LargestConnectedComponent(shapes,dims,skipBias):
     """
     L=len(shapes)-skipBias     
     shapes=shapes.reshape((-1,) + dims[1:])    
-    structure=np.ones(tuple(3*np.ones((np.ndim(shapes)-1,1))))
+    structure=np.ones(tuple((3*np.ones((np.ndim(shapes)-1,1))).astype('uint')))
     for ll in range(L): 
         temp=np.copy(shapes[ll])
         CC,num_CC=label(temp,structure)
